+++ Multifabriken/produkter.cs
@@ -1,0 +1,11 @@
+namespace Multifabriken
+{
+    abstract class produkter
+    {
+        public abstract void angeData();
+
+        public abstract void addToList();
+
+        public abstract void presentList();
+    }
+}
